Run `did-update` callback in an untracked frame
Modifier capabillities in 3.22, requires all arguments to be consumed,
to be tracked. This meant `disableAutoTracking` had to be set to `false`
for the modifier to work with 3.22 capabillities. This had the
sideeffect that any other consumed tags ended up being tracked. Meaning
the `did-update` callback was invoked whenever those tags changed.